+++ routes/organization.js
@@ -15,6 +15,11 @@ router.get(
   Security.checkCapability("organization", ["view"]),
   organizationController.getAllOrganization
 );
+router.get(
+  "/:organizationId/invite",
+  Security.checkCapability("organization", ["view"]),
+  organizationController.getOrganizationInvites
+);
 router.post(
   "/",
   Security.checkCapability("organization", ["add"]),
@@ -29,6 +34,16 @@ router.post(
   "/:organizationId/invite",
   Security.checkCapability("organization", ["edit"]),
   organizationController.putInviteToOrganization
+);
+router.post(
+  "/:organizationId/invite/:inviteId/resend",
+  Security.checkCapability("organization", ["edit"]),
+  organizationController.postResendOrganizationInvite
+);
+router.delete(
+  "/:organizationId/invite/:inviteId",
+  Security.checkCapability("organization", ["edit"]),
+  organizationController.deleteOrganizationInvite
 );
 router.delete(
   "/:organizationId",